Added /h globalstats, added LaunchProjectileAction, added new way to set materials with enums, idk what else - I wanted to get more done today, but spent awhile getting the Redstone Studios House in order, there is a lot of drama going on there :) - Its fine though got some stuff done on my todo list :) - Made function automatic executions more consistent

+++ src/main/java/com/al3x/housing2/Instances/HousingData/FunctionData.kt
@@ -9,7 +9,7 @@ data class FunctionData(
     val name: String,
     val id: String,
     val description: String,
-    val ticks: Double?,
+    val ticks: Int?,
     val material: String,
     val global: Boolean = false,
     val actions: List<ActionData>
